src/service: return D-Bus error if slot states cannot be determined
For now the service silently ignored if it could not determine slot
states. This led to a D-Bus response with a status message that
did not have the slots "state" member set to a proper value.

The CLI trying to interpret this then ends up in the undefined state of
r_slot_slotstate_to_str() that emits a g_error()

| (rauc:1523868): rauc-ERROR **: 12:32:50.829: invalid slot status 0

Fix this by not ignoring the error but forwarding it through D-Bus so
that the CLI or other frontends do not try to interpret broken data.

For this we equip create_slotstatus_array() with proper GError handling
and slightly adapt the flow in r_on_handle_get_slot_status() in order to
properly handle this and forward the error as a D-Bus error using
g_dbus_method_invocation_return_gerror().

While an error in determine_boot_states() cannot be fatal, at least
increase the log level from 'debug' to 'message'.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>