bundle: fix unsquashfs arguments and extend test
Remove the incorrect '-e' argument to unsquashfs, which is used to pass a file
containing filenames to extract. As unsquashfs parses all arguments after the
filesystem image path (the bundle in our case) as 'directories or files to
extract', the '-e' didn't have any effect unless a bundle contained a file
named '-e'.

Also add some tests that we only extract the expected file from the bundle.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>