.github: workflows: compile tests on ubuntu stable releases
In the past, we did not always notice when breaking RAUC compatibility
with older toolchains and libraries.
While some incompatibilities cannot and will not be avoided (like
dropping support for OpenSSL 1.0.x that is EOL since end of 2019) others
are not strictly required and may happen unintentionally.

These added test should allow to be notified when breaking compatibility
with old releases and making the decision explicit rather than implicit.

The tests currently run for ubuntu releases:
  - 16.04 (gcc 5.3.1, linux-libc-dev 4.4.0-197.229, glib 2.48.2, openssl 1.0.2g)
  - 18.04 (gcc 7.3.0, linux-libc-dev 4.15.0-20.21, glib 2.56.4, openssl 1.1.1)
  - 20.04 (gcc 9.3.0, linux-libc-dev 5.4.0-59.65, glib 2.64.3, openssl 1.1.1f)

Note that for now they test only compilation. Not runtime behavior.

As it might happen, that support for a component is dropped
intentionally, we add the ability to place specific non-mainline
backport patches for each ubuntu version.

Patches currently required for 16.04 are:

- Re-Add support for OpenSSL 1.0.2
- Do not use glib macro G_VARIANT_DICT_INIT which is not available in
  glib 2.48.2

Building against 16.04 also requires to disable GPT support as this
needs fdisk at least in version 2.29, but only 2.27 is available.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>