configure.ac: check for specific GLIB min/max API version
In the past we had some issues with unattended increases of the required
glib API version due to the use of methods that were introduced in newer
version than the one we currently explicitly depend on in
PKG_CHECK_MODULES.

To prevent from unintentionally using new API methods, glib provides the
GLIB_VERSION_MAX_ALLOWED macro the can be set to the maximum allowed API
version prior to including glib.h.

When using methods that belong to a more recent glib API, this will emit
a compiler warning like

| src/dm.c:30:2: warning: ‘g_uuid_string_random’ is deprecated: Not available before 2.52 [-Wdeprecated-declarations]
|    30 |  dm_verity->uuid = g_uuid_string_random();
|       |  ^~~~~~~~~

The workflow idea is to give the developer a tool to be notified about
using a newer API and moving the need to use it from an implicite to an
explicite decision. If the usage gives some major benefits, the
developer can decide to

a) strictly depend on a newer glib and update
   - required version in PKG_CHECK_MODULES
   - GLIB_VERSION_MIN_REQUIRED
   - GLIB_VERSION_MAX_ALLOWED

b) add a fallback to compile against the curently supported glib API
   - update GLIB_VERSION_MAX_ALLOWED
   (use carefully as this will prevent further warnings for new methods
   of same API version)

Note that it is required to also explicitly set
GLIB_VERSION_MIN_REQUIRED as it is pre-defined to be
GLIB_VERSION_CUR_STABLE which can be even newer than
GLIB_VERSION_MAX_ALLOWED.
Since we might check fo API macros not being defined in older glib
versions, we use G_ENCODE_VERSION(major,minor) rather than the
pre-defined GLIB_VERSION_major_minor macros.

The initial values reflect the current state of API usage:
- The maximum version is 2.52 as we use g_uuid_string_random() method
  from 2.52 API
- The minimum required version is 2.50 since we have a fallback
  defintion for g_uuid_string_random() but use G_VARIANT_DICT_INIT macro
  that was introduced with the 2.50 API

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>