src/bundle: fix casync manifest generation
create_bundle() calls update_manifest_checksums(), which would result in
checksums and sizes that refer to the casync index files instead of the actual
image. While this only breaks installation when using UBI volumes with casync
(which needs an additional casync patch), the incorrect checksums are still
confusing in the slot status information.

As we already have saved the correctly updated manifest, just call mksquashfs()
and sign_bundle() directly.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>