fix too strict payload size check
RAUC versions up to 1.4 had an issue in the casync bundle signature generation,
which caused two signatures to be appended. While the squashfs payload size is
a multiple of 4KiB, the end of the first signature was not aligned. As RAUC
uses the second ("outer") signature during verification, this didn't cause
problems. RAUC 1.5 fixed the casync bundle generation and added stricter
checks, which rejected the older bundles.

To allow installing old casync bundles, we reduce the message to a notification
and point to the explanation in the FAQ.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>