src/update_handler: run boot-mbr-switch post install hook prior to MBR switch
RAUC's update logic works by performing all necessary updates and then
switching the active slot in an atomic operation. Apply the same logic
to "boot-mbr-switch" slots by running the slot post install hook prior
to switching the active/inactive partition in the MBR.

This allows manipulation of written images expecting an absolute offset
on the raw mmc device rather than an offset to the partition start. This
applies to the i.MX28 BCB (boot control block) header.

Extra information for the hook to manipulate data on the not yet
activated partition will be passed in a future commit.

Signed-off-by: Bastian Krause <bst@pengutronix.de>