src/update_handler: run boot-gpt-switch post install hook prio to GPT switch
RAUC's update logic works by performing all necessary updates and then
switching the active slot in an atomic operation. Apply the same logic
to "boot-gpt-switch" slots by running the slot post install hook prior
to switching the active/inactive partition in the GPT.

Similar to "boot-mbr-switch", this allows manipulation of written
images.

Extra information for the hook to manipulate data on the not yet
activated partition will be passed in a future commit.

Signed-off-by: Bastian Krause <bst@pengutronix.de>