src/update_handler: run boot-emmc post install hook prio to r/o and active partition switch
RAUC's update logic works by performing all necessary updates and then
switching the active slot in an atomic operation. Apply the same logic
to "boot-emmc" slots by running the slot post install hook prior
to switching to r/o and switching the active/inactive partition.

Similar to "boot-mbr-switch" and "boot-gpt-switch", this allows
manipulation of written images.

Extra information for the hook to manipulate data on the not yet
activated partition will be passed in a future commit.

Signed-off-by: Bastian Krause <bst@pengutronix.de>