Fix configurable U-Boot attempt setting
Commit ebbdb4be9d ("Make U-Boot attempt settings configurable") made
it possible to tweak the number of attempts that are set in the U-Boot
environment.

There is an issue with that because U-Boot's setexpr can only handle
hex, while in the current implementation those values are treated as
decimals.

To work around that, let's treat those numbers as hex everywhere.

Signed-off-by: Pascal Huerst <pascal.huerst@gmail.com>