src/manifest: treat unknown hook names as errors
In the past, we had several cases where people misspelled hook names when
writing the manifest file, like

  [hooks]
  filename=hooks.sh
  hooks=invalidname

As this caused only warnings when generating the bunde on host side, the
misconfiguration kept hidden until trying to install the bundle on the
target.

With turning the warning into an error, the bundle creation will now
explicitly fail and report the invalid hook configuration.

Note: This could cause compatibility issues in the very rare case that
one attempts to use an old RAUC host version to create a bundle for a
newer RAUC target version which got a new hook defined.
For this case, we strongly encourage to simply update the host RAUC
version, too.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>