src/manifest: check manifest filenames against bundle content dir
This introduced sync_manifest_with_contentdir() that not only updates
the manifest checksums based on the filenames provided, but also checks
for the image files and hook file being present in the bundle content
dir.

While the presence of the image file was checked indirectly by
update_manifest_checksums(), the hook file was not checked at all.

This now adds dedicated checks with helpful error messages for both
cases so that bundle creation should be easier to debug.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>