src: allow passing separate devicetree file to barebox-state
This makes use of the new feature of dt-utils' barebox-state command
'-i' switch that takes the path to a devicetree file to read the state
information from (introduced by v2021.03.0 release).

We add a new configuration option only being enabled when using
'barebox' boot selection backend: `barebox-dtbpath`.
Example:

  [system]
  ...
  bootloader=barebox
  barebox-dtbpath=/run/efi/EFI/BAREBOX/state.dtb

This is especially meant to be used for platforms that do not use
devicetrees for booting, thus mainly x86 ones.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>