src/manifest: remove parsing and writing of obsolete 'keyring' section
The keyring section is a remnant from the initial implementation of RAUC
but was only parsed as a key but never used anywhere, actually

Following the general concept that a RAUC manifest / bundle should
exactly describe the installation, an uninterpreted key is considered to
be an error and should not be silently accepted.

Thus, despite this might potentially break very rare cases where users
accidentally placed a 'keyring' section in their manifest, remove this
entirely to let the installation fail explicitly (and with a meaningful
error message that an unsupported section was used).

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>