src: remove all remnants of long deprecated file support
The 'file.<slotclass>/filename' sections are a remnant from the old
non-bundle network mode that was removed a while ago already.
They had a similar purpose as 'image' sections for bundle mode have.

Following the general concept that a RAUC manifest / bundle should
exactly describe the installation, an uninterpreted key is considered to
be an error and should not be silently accepted.

Thus, despite this might potentially break cases where users
accidentally placed file sections in their manifest, remove this
entirely to let the installation fail explicitly (and with a meaningful
error message that an unsupported section was used).

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>