src: support omitting filename when 'install' slot hook is used
The 'install' slot hook may potentially change the slots content without
strictly requiring an image. Thus we should skip checking for an image.

As we checking (or assuming) that image->filename is not NULL at several
places in the code, this requires some changes across the install and
manifest parsing module.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>