src/main: limit --override-boot-slot argument to specific subcommands
For most commands, --override-boot-slot will not have any impact as this
is only evaluated for the service (or 'install' command when RAUC is
being compiled without 'service' support).

Fix this by moving the parameter from the global option list to the
option list of 'service' (and 'install' / 'status').

Also fix the sharness tests that had this argument set without any
effect.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>