src/manifest: support older bundles created with zero size images
Up to RAUC 1.5, RAUC bundle generation allowed to use images with zero
size. This could have been used as a hack to prevent from having to
provide an image file for an 'install' per-slot hook that, e.g., copies
data from the active to the inactive slot.

As RAUC did not differentiate between no size set and a size of zero,
the generated manifest did not contain a 'size=' parameter for the
respective 'image' section, e.g:

   [image.datafs]
   filename=e3b0c4...
   sha256=xxx
   hooks=install;

To keep compatibility for these kind of bundles, make the check for
missing 'size=' parameter non-fatal when an 'install' hook is used.
Instead, only prints an info message and fall back to the former behavior
of assuming size being '0' (otherwise subsequent update handlers will
fail when trying to write an image of size '-1').

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>