src/bundle: fix error handling for small squashfs sizes
When having a squashfs that is only 4096 bytes in size, we cannot
generate a hash table as verity would directly calculate the root hash.

To catch these issues, an assert was places in the code.
However, the assert must check for '>' instead of '>=' as otherwise it
would allow 'combined_size*4096' equal to 'offset' which results in a
'verity_size' of 0.

As the size of 'offset' actually depends on user input (i.e. bundle
content size), we should not use g_assert() as primary check instance
here, but add a GError check that allows to emit a user-readable error.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>