src/update_handler: support oe-core's default SquashFS image suffixes
The Yocto/OE 'image_types' class generates SquashFS images with different file
name suffixes, depending on the compression selected (e.g. .squashfs-xz,
.squashfs-lzo, ...). With our current image pattern, this would require an
extra renaming step (to .squashfs) to allow RAUC to properly recognize the
image as a SquashFS.

To avoid matching images by mistake (which contain .squashfs somewhere else in
the filename), use an explicit list.

As Yocto/OE is probably one of the most common build ecosystems RAUC is
currently used with, this appears to be an acceptable compromise.

Signed-off-by: Omer Akram <omer@thing.com>
Signed-off-by: Jan Luebbe <jlu@pengutronix.de>