feature: add option to create an unsigned bundle
Many industrialization procedure requires to sign any artifact in a
dedicated secure room with restricted access (as a PKI).

This proposal add a `--external-signing` option to bundle command which
generate an intermediary artifact to sign externally.

The result for the respective formats:

- `plain` produces an unsigned raw bundle
- `verity` produces also an unsigned raw bundle but additionally creates
  a RAUC `verity` manifest named `manifest-dm.raucm` in the same
  directory than the unsigned raw bundle

Any suggestion are welcome. In case of interrest I will propose
secondarily a command to create a bundle from external signature and the
associated tests (in addition to the tests for this proposal).

Signed-off-by: Jean-Pierre Geslin <jarsop@outlook.com>