src/bootchooser: introduce r_boot_is_supported_bootloader()
The information which bootloader is supported should be provided by
the bootchooser module that also implements the support, not by the
config_file module that is responsible only for parsing.

Thus we move the definition of supported bootloaders there and
encapsulate checking implementation by a transparent checker method
r_boot_is_supported_bootloader().

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>