src/bundle: Allow bundle to be stored on ramfs filesystem
If the bundle being checked by `check_bundle_access()` is stored on a
filesystem of type 'ramfs' it will fail with the error message "unable
to find mounted device for bundle".  Fix it by treating the 'ramfs'
filesystem type the same as 'tmpfs' ("prepared by root" so no need to
check the underlying device, which does not exist anyway).

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>