src/update_handler: adapt MBR/GPT boot partition switch terminology
In dbd8f2df ('docs/advanced: clarify how boot-mbr-switch works') and
27c0bdf0 ('docs/advanced: update boot-gpt-switch explanation to
"first/second half" wording') the documentation for MBR/GPT boot partition
switching was clarified and made more precise.

However, the log output for these functions was not adapted and is thus
misleading/wrong.

This is fixed here to reflect that we do not switch between different
partitions but switch the location of the boot partition between the
first and the second half of a defined region.

Also, adapt the variable name 'inactive_part' to 'inactive_half' for
this.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>