src/main: allow passing 'no-verify' to rauc convert
This option allows to disable the verification of the input bundle (and
thus omitting --keyring argument) when converting a conventional to a
casync-based bundle.

However, this is not recommended as signature checks are there for a
reason of course.

But, as we also already allow this for similar subcommands like ('info' or
'resign') and should give the responsible user the freedom to 'know what
he's doing', we enable this for convert, too.

Another quite relevant point where this can be used as a workaround is
when being inside an environment (e.g. a docker container) where the
standard RAUC bundle access check mechanisms must fail) and trigger
errors like

| initial check_bundle_access failed with: unable to find mounted device for bundle

This happens when RAUC attempts to find the underlying device of the
mountpoint the bundle is located at but cannot as the device is hidden
by the container.

Addresses: #679, #688

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>