src: introduce --trust-environment option for 'rauc convert'
In some cases it might not be possible or wanted that RAUC (used as a host
tool) performs its standard bundle access checks.

For example when running 'rauc convert' from within a (docker)
container, RAUC might not be able to get the underlying device of the
bundle's storage location. This way it cannot determine if the mount
point is owned by root and will abort with

| initial check_bundle_access failed with: unable to find mounted device for bundle

While --no-verify potentially already works around this, it is a too big
hammer for most cases where you actually want to benefit from the usual
signature verification of the input bundle and just get stopped by the
exclusive access check.

Now the --trust-environment option allows to tell RAUC that we assume we
have our environment we run RAUC in under control and it can skip the
exclusive access checks.

Fixes #679, #688

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>
Signed-off-by: Jan Luebbe <jlu@pengutronix.de>