src/update_handler: if seed slot is mounted ro, remount it ro
Re-mounting a ro (readonly) slot rw (readwrite) will cause an error or a
fallback with a warning, depending on the mount version/veriant used.

Also mounting an rw slot as ro will fail.

Thus make sure we use the same option when remounting the slot for
seeding that the original mount already had.

Fixes #773

Signed-off-by: Jonas Licht <jonas.licht@gmail.com>
Signed-off-by: Enrico Jorns <ejo@pengutronix.de>