utils: Introduce UBI device name resolving
The U-Boot distroboot cmd mechanism and maybe others use the alternative
ubi rootfs format with kernel cmdline parameter 'root=ubi0:volname'.
The kernel resolves this to a ubi volume and mounts it as rootfs.  This
is nice, because it does not contain mtd nor ubi device numbers.  The
UBI userland tools and RAUC however only work with paths like
'/dev/ubi1_3' which map to the UBI volumes known to the system.

To resolve the alternative format to a device path, the same parser
logic as in the kernel itself is used, and all the needed information to
construct the device path is collected from sysfs.

This can be used to let RAUC support the alternative format in both
kernel cmdline parsing and config file, which would make it easier to
integrate with U-Boot and easier to configure by the administrator,
because volume names with a meaning can be used instead of hardcoded
device paths with numbers.

Signed-off-by: Alexander Dahl <ada@thorsis.com>