src/install.c: do not emit unreliable slot states
When not having the booted slot detected and not being in a special case
like nfs boot or rauc.external, we should not set any slot states at
all.

Up to now, the function correctly emitted an error but fully initialized
the slot states with unreliable values (treating all as 'inactive' to be
concrete).
The calling code in service.c:create_slot_status_array() then dropped
the error by printing it with a g_debug() and used the returned slot
states carelessly.

We can simply fix this by moving the error check for '!booted' above the
actual slot state assignment. This ensures the slot states are
initialized only when the function succeeds.

Additionally, in case of an external boot, we explicitly need to set all
slots to 'inactive'.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>