src/bootchooser: really remove redundant error message
Unlike the first attempt (a14509f2) to remove the following redundant message:

| Failed getting primary slot: Failed getting primary slot: Unable to obtain primary slot

this approach now fixes all occurences and provides a more consistent
prefixing scope;
A function must only prefix an error message if it can add information
an ouside caller cannot not add.

Example:
The call of make_me_a_sandwhich() does not need to return error messages
with a prefix "Cannot make a sandwich: " as the caller knows that this
function could not make a sandwich in case of an error and thus the
caller could add the information itself.
However, the function make_me_a_sandwich() could add an error prefix "No
cheese found: " as this is information is not available in the outside
scope.

Without stricter rules on error prefixing we might come across this
issues (and thus double-prefixed error messages) again and again.

Fixes: a14509f2 ("src/main: remove redundant "Failed getting primary slot:" error string")

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>