src/bootchooser: custom_get_primary: provide helpful error message
The fact that the primary slot could not be found is already indicated
by the error returned from the method and does not add any useful value.

Replace this by the quite more useful information that the bootname that
was returned by the 'get-primary' call does not match any of the known
bootnames of any slot.

A 'rauc status' call error message for example then turns from:

| Failed getting primary slot: Unable to obtain primary slot

into

| Failed getting primary slot: 'foobar' does not match any bootname

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>