src/install: determine_boot_states: do not abort on first failed slot
The determine_boot_states() method iterates over (boot) slots to determine
their boot status.
If determining the boot status failed for one slot, the method aborted
instantly and determine_boot_states() returned an error.
With this handling, it was kind of random which boot status was set and
which not, because this actually depended on the order the slot were
tested and which slot in this order returned an error.

As a result, even slots for which the boot state could have been
determined as 'good', remained marked 'bad'.

To fix this behavior, do not abort on any failed slot but just print the
error message and continue.
Just save the information that not all slots could be determined
sucessfully and report an error at the end.

This is even more consistent when having in mind that the error of
determine_boot_states() is not handled at all but only emitted in a
g_debug() message. Thus higher layers are not informed why the
respective slot boot state is bad.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>