config_file: Resolve special device names
When we allow using special device names like ubi0:rootfs1 as parameter
for 'root=' in kernel cmdline, it makes sense to use the same resolver
mechanism for the config file.  This makes the config more
understandable from an administrator point of view and less error prone,
because volume names with a meaning can be used instead of unstable
device paths which might change depending on volume creation order.
Example for a slot configuration:

    [slot.rootfs.1]
    device=ubi0:rootfs1
    type=ubifs
    bootname=A

    [slot.rootfs.2]
    device=ubi0:rootfs2
    type=ubifs
    bootname=B

Signed-off-by: Alexander Dahl <ada@thorsis.com>