Add casync test/update_hander.c test ignore-seed-slot
Test was added while trying to follow the same patterns as
in rest of test cases located in the same file. Due to the fact
that casync is already missing the base for testing, this one
is relatively large for what it actually tests.

As it is, the test does alot of setup and tests very little.
But the test itself goes as follows:

* Run test with default casync configuration (ignore-slot-seed=false)
* * Test sure that casync program is called WITH --seed option
* Run test with ignore-slot-seed=true
* * Test that casync program is called WITHOUT --seed option

Signed-off-by: Esa Laakso <esa.laakso91@gmail.com>