Add config ignore-slot-seed and add test config_file_test_casync_options
ignore-slot-seed is required in configuration file. Add it to config -structure
to be configurable, with reading in load_config.

Add test to check that the new configuration option is read, and that it has
a default value (false) when not specified. Casync options in general are
missing a test in config_file.test.

Add a bare minimum test to check that other options (storepath and tmppath)
are also read in some way.

ignore-slot-seed flag checks in installation are based on code made by
fhackenerger, located in PR https://github.com/rauc/rauc/pull/513, with
excess code refactored/removed to fit into config option.

Signed-off-by: Esa Laakso <esa.laakso91@gmail.com>