qemu-test: remove trailing comma from nandsim.parts commandline parameter
nandsim.parts parameter is a comma-separated array, but should not be
terminated with a comma.
Otherwise kernel option parser will ignore the parameter with:

| Booting kernel: `0x100' invalid for parameter `nandsim.parts'

However, this is not critical as 0x100 is equal to the default partition
size erase.

Fixes 318640ca ("qemu-test: add setup to test MTD and block devices")

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>