.github: workflows: move 'distcheck' to separate workflow
Currently, the 'tests' workflow fails almost every time because the 'Run
distcheck' step triggers a test failure.

The underlying failure is not RAUC-specific and related to loodev mount
and cache handling.

Details can be found in https://github.com/rauc/rauc/issues/790

As fixing this issue is not straight forward and the relevance for
real-world scenarios is hardly given, a workaround needs to be
established.

Tests showed, that the distcheck loop mount failure is not triggered (or
significantly less often) when running directly in the GitHub action runner vm
instance (rather than in the rauc testing container).

This commit thus moves the error-prone distcheck from the 'tests'
workflow to a separate 'distcheck' workflow.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>