src/update_handler: add support for jffs2 tar archives
This patch introduces support for tar archives for jffs2 flash storage.

For jffs2 we have to use the MTD chardev in order to erase the flash
storage. The subsequent mounting call has the peculiarity of needing a
different device parameter than a device path, meaning that no path must
be passed. We take care of that with specific handling in mount.c

Signed-off-by: Holger Assmann <h.assmann@pengutronix.de>