src/bundle: do not check for bundle format identifier anymore
For local bundles, we used the squashfs marker to check for the bundle
being valid. This will not apply to encrypted bundles anymore (and was
not a sufficient indicator anyway and not used for remote bundles).

Instead, use exisiting check for detached CMS to detect valid CMS data
and make error reporting more explicit and verbose.
While at it, also be more careful with BIO mem allocation in cms_is_detached().

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>