.github: workflows: drop support for Ubuntu 16.04 (Xenial)
Xenial is out of standard support since April 2021. RAUC itself dropped
official support for OpenSSL < 1.1.1 with v1.3 (released on 2020-04-23). For
RAUC v1.5, we had a backport patch for OpenSSL 1.0.2 to make it easier to apply
the security fix.

By now, updating this backport patch for new RAUC changes is no longer
appropriate, as RAUC v1.5 can be used as an intermediate step while moving to a
newer base system.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>