src/context: optionally fall back to loading default config path
In 629a7c1, config file loading was reworked to allow printing explicit
failure messages when a config is required (or given) but not loadable.

Functionality that was lost with this change is the ability to
optionally use a system config file if it exists under the default path.

This primarily hits your when attempting to run `rauc info` on the
target. There it implicitly loaded the keyring from
/etc/rauc/system.conf if this existed. With the change noted above, one
would have needed an explicit --keyring argument then.

To fix this, always try to load a config from the default configuration
path /etc/rauc/system.conf before defaulting to the built-in
default_config().
If attempting to load the configuration should return another failure
than 'No such file or directory', again propagate this to the caller as
we still do not want to silently drop broken configs.

Note that explicit setting of `context->configpath` is still used to
enforce the availablity of the config file.

Fixes 629a7c1 ("src: require config loading to succeed if config path is configured")

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>