src/bundle: fix rauc convert for empty filename
Commit 8a9c9213 introduced the ability to not provide any image file at all in
case an 'install' slot hook is used.

But, when trying to convert such a bundle (to a casync bundle), the
method convert_to_casync_bundle() will call image_is_archive() to
determine if a blob or a directory index needs to be created.

So far, this method silently assumed that `image->filename` is always
set, which it is not anymore.

Fix this by explicitly skipping image entries with no 'filename' set and
additionally add g_return_val_if_fail() guards in image_is_archive() to
prevent from using it wrongly.

Fixes: 8a9c9213 ("src: support omitting filename when 'install' slot hook is used")

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>