src/context: always set context configpath when config is used
Code expects context's configpath to be set if config file is required.

This is an issue for example during installation when it is used to set
the environment variable RAUC_SYSTEM_CONFIG for pre-install handlers.
With current handling, as introduced by a61bb4e0, this causes the rauc
service to die with:

| g_environ_setenv: assertion 'value != NULL' failed

For consistency, simply always set the configpath variable when a config
file was loaded from any location.

Fixes: a61bb4e0 ("main: specifiy configuration mode per command")

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>