src/main: expose --cert and --key arguments for signing commands only
It is not necessary (but instead misleading) to have the --cert and
--key arguments as global options because they effectively apply to only
a very few commands.

Introduce a purpose-driven (rather than subcommand-driven) GOptionEntry
for this.

For backwards-compatibility of existing 'rauc {bundle, resign,
convert}' command lines calls, we keep the original global options, but
add the G_OPTION_FLAG_HIDDEN flag to not have them in the info output.

For the command-specific option group we must set G_OPTION_FLAG_NOALIAS
flag to turn of glib's automatic conflict resolution for arguments that
would prefix the arguments with 'groupname-'.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>