manifest: add support for a list of optional incremental methods
These methods will add extra information to the bundle, allowing RAUC to
access only the parts of an image which are not yet available locally.
Together with streaming, this reduces the amount of downloaded data.

As the full image is still available in the bundle, older RAUC versions can
ignore unsupported incremental methods and we don't need to check the list for
explicit values. Currently, no methods are implemented.

Note that because we reject any manifest with unknown keys, bundles
which use incremental methods will be rejected by versions up to v1.6. By
integrating this change early (without actually implementing any methods), we
can add new methods that will be ignored by older versions. This way, users
can generate bundles with incremental methods without compatibility problems as
soon as a version with this minimal change is deployed on all devices.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>