add initial support for desync
Desync is an alternative casync implementation.
Unfortunately it's not entirely a drop in replacement because they
differ for some of their launch options.

This commit adds the `use-desync` config key that offers the ability to
use desync instead of casync.

Currently the only downside is that it doesn't use the active partition
as a seed.
This is expected to be added with a future commit, but in the meantime
the `install-args` configuration option could be used to manually
specify the seeds that need to be used.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>