src/mount: prevent double-close in r_setup_loop()
In the for loop, we call g_close() on loopfd. If the following ioctl
fails, we leave the loop with loopfd closed but being >= 0.

As a consequence, the g_closer() after the 'out:' label will again try
to call g_close() on loopfd, resulting in a double-close.

To prevent from doing this, simply invalidate loopfd after having closed
it.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>