src/bundle: don't let existing bundles disappear
The cleanup logic meant to remove incompletely generated
artifacts/bundles on error is wrong and thus also removes existing
files under bad circumstances.

The current logic is based on always removing the generate output in
case of an error, except when the file exists and the error was a
G_FILE_ERROR_EXIST error.

The G_FILE_ERROR_EXIST error is thrown by explicit file existence checks
in truncate_bundle(), mksquashfs() and convert_to_casync() functions.

The problem is, that these methods are not the first ones run in the
respective output-generating methods. Most run check_bundle_payload()
first for example. And if this fails, and the output path already
existed, then this will just be silently removed.

An simple example of where this will hit you is when running:

  rauc bundle [..] content-dir out.raucb
  echo "[invalidsection]" >> content-dir/manifest.raucm
  rauc bundle [..] content-dir out.raucb

The first call will create 'out.raucb' while the second will remove it.
Here the reason is that load_manifest_file() fails (due to the incorrect
section name).

Resolve this by explicitly early-checking for output file existence in
every bundle-generating function and returning instantly without
attempting to remove the output file.

We can then safely remove or simplify the existing checks.

Also note that in a more detailed look, there have been more issues.
In encrypt_bundle() for example even the g_error_matches() check was
missing and thus the bundle removed on each and every error.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>