src/main: Add --trust-environment for 'rauc extract'
This option was already available for other commands, like
'rauc convert'.

Sometimes a bundle needs to be extracted but it may not be possible, or
desirable, to perform the standard bundle access checks.

For example if RAUC is executed from a controlled environment where the
ownership of a bundle is not an issue. Or when obtaining the underlying
device is not possible at all, e.g. when inside an unprivileged Docker
container.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>