src/update_handler: extend copy_raw_image() to write a header at last
There are some ROM loaders that look for an image header and/or magic
numbers to determine if a valid image is programmed at a particular
address. If the image is linearly programmed into memory, starting with
the header, the loader may mistakenly treat the memory location as valid
in the event of an interrupted update. To avoid such scenarios, the header
must be written last.
Add an option to copy_raw_image() to write a header section at last.

Signed-off-by: Matthias Fend <matthias.fend@emfend.at>