manifest: ignore meta.<label> sections in the manifest
The meta.<label> sections are intended to provide a forwards-compatible way to
add data to the manifest which is not interpreted by RAUC in any way.
Currently, they are just ignored when reading a manifest. In future releases,
they will be accessible via rauc info, the D-Bus API and in hooks/handlers.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>