src/bootchooser: stricter error message scoping when marking good/bad
On the bootchooser backend level, setting boot status is not referred to
as 'marking'. If we apply the rule to not add redundant information, we
can and should simply shift this printout to mark.c.

This also fixes redundant printout when r_boot_set_state() is not
invoked from manual marking but from installation process which already
added "Failed marking slot %s non-bootable: %s" as prefix and produced
error messages like

| Failed marking slot rootfs.1 non-bootable: Failed marking 'rootfs.1' as bad: ...

which now show as

| Failed marking slot rootfs.1 non-bootable: ...

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>