test/rauc.t: do not operate on original install-content and cleanup
The test/install-content/ directory should be seen as read-only.
Otherwise tests can interfere as they modify directory content in-place.

For 'make distcheck' this even solves significant test issues since here
the install-content in build dir consists of symlinks only. These are
not resolved when calling 'rauc bundle' on it.

While this does not make tests fail in most cases, it leads to errors
for verity bundles that require a minimum size and also leads to
problems with 'rauc convert' that tries to access broken symlinks in
bundles.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>