src/rauc.Installer.xml: Add new InfoBundle method
The old method lacked the capability of passing additional
arguments, which prevented from adding a lot of features in the past.

Compared to the existing one, the new method differs only in the
additional array-of-string-variant parameter 'args'.

This allows to pass all kind of future additional simple arguments.

To guarantee API stability, the old method will be kept as an
alternative.

Signed-off-by: Stefan Ursella <stefan.ursella@wolfvision.net>