src/service: implement callback handling for InspectBundle
With the old 'Info' API it was not possible to pass HTTP authentication
options for RAUC's streaming bundle access.

The new API now also removes another major limitation of the old one:
Return data can now be extended without API breakage since bundle
information is returned as an 'a{sv}' dict representations.

For now, extend this to return the the manifest 'update' section with
the four base information fields 'compatible', 'version', 'description',
and 'build'

Signed-off-by: Stefan Ursella <stefan.ursella@wolfvision.net>
(reworked to match new name and generic a{sv} return type)
Signed-off-by: Enrico Joerns <ejo@pengutronix.de>